example/client/requirements.txt
grpcio>=1.34.0
aiohttp>=3.7.0
tensorflow>=2.3.0
tensorflow-serving-api>=2.3.0
